+++ lib/main.dart
@@ -1,15 +1,16 @@
-import 'package:app/announce_screen.dart';
-import 'package:app/group_screen.dart';
-import 'package:app/login_screen.dart';
-import 'package:app/message_screen.dart';
-import 'package:app/register_user.dart';
+import 'package:app/screens/announce_screen.dart';
+import 'package:app/screens/group_screen.dart';
+import 'package:app/screens/login_screen.dart';
+import 'package:app/screens/message_screen.dart';
+import 'package:app/screens/register_user.dart';
+import 'package:app/screens/splash_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'group_created_screen.dart';
-import 'home_screen.dart';
-import 'join_group.dart';
-import 'main_screen.dart';
-import 'create_group_screen.dart';
+import 'screens/group_created_screen.dart';
+import 'screens/home_screen.dart';
+import 'screens/join_group.dart';
+import 'screens/main_screen.dart';
+import 'screens/create_group_screen.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -25,11 +26,11 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       initialRoute: '/',
       routes: {
-        '/': (context) => const MainScreen(),
+        '/': (context) => SplashScreen(),
+        '/mainscreen':(context) => const MainScreen(),
         '/welcome': (context) => const LoginScreen(),
         '/home': (context) => const HomeScreen(), // Add a route for HomeScreen
-        '/createGroup': (context) =>
-            CreateGroupScreen(), // Add Create Group screen route
+        '/createGroup': (context) => CreateGroupScreen(), // Add Create Group screen route
           //'/groupCreated': (context) => GroupCreatedScreen(), 
           '/Registration': (context) => const Registration(),
           '/joinGroup':(context) => JoinGroupScreen(),
