+++ lib/screens/join_group.dart
@@ -2,7 +2,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'UserInfo/joiningGroupmodel.dart';
+import '../UserInfo/joiningGroupmodel.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 class JoinGroupScreen extends StatelessWidget {
